Minor fixes

+++ board_server/Src/usart.c
@@ -134,7 +134,7 @@ void HAL_UART_MspDeInit(UART_HandleTypeDef* uartHandle)
 
 /* USER CODE BEGIN 1 */
 int write_on_uart2(char * str){
-	return HAL_UART_Transmit( &huart2, ( uint8_t * )str, strlen( str ), 5000 );
+	return HAL_UART_Transmit( &huart2, ( uint8_t * )str, (uint16_t) strlen( str ), 5000 );
 }
 /* USER CODE END 1 */
